Fix all BacklogService performance and correctness issues
1. findPersisted() — O(1) via #Predicate + fetchLimit:1
   Was: fetch ALL records, linear scan in memory.
   Now: SwiftData predicate filters at the store level.

2. No more fetchTasks() after every save
   Was: every mutation did full re-fetch + map of all records.
   Now: local mutations update the in-memory `tasks` array directly.
   Full re-fetch only on remote CloudKit changes.

3. Error handling for SwiftData saves
   Was: `try? modelContext.save()` — errors silently swallowed.
   Now: `saveContext() -> Bool` with observable `lastError` property.
   Mutations guard on save success — no in-memory update on failure.

4. Fractional sortOrder (Double) for conflict-safe reordering
   Was: sequential Int — a reorder renumbered ALL records, causing
   CloudKit conflicts when two devices reorder simultaneously.
   Now: Double with midpoint calculation. A reorder only touches ONE
   record. Automatic renumber when gap < 1e-6 (~50+ halvings).

5. lastRemoteUpdate for CloudKit sync visibility
   Observable Date? set when NSPersistentStoreRemoteChange fires.
   UI can show "synced just now" indicator.

6. carryOverUnfinished uses predicate
   Only fetches scheduled tasks instead of all tasks.

+++ Bubo/Models/PersistedBacklogTask.swift
@@ -24,12 +24,15 @@ final class PersistedBacklogTask {
     var modifiedAt: Date?
     var scheduledDate: Date?
     var scheduledEventId: String?
-    /// User-controlled ordering (lower = earlier in list).
-    var sortOrder: Int = 0
+
+    /// User-controlled ordering.  Uses `Double` with gaps (fractional
+    /// indexing) so that a reorder only touches ONE record — minimises
+    /// CloudKit conflicts when two devices reorder simultaneously.
+    var sortOrder: Double = 0
 
     init() {}
 
-    init(from task: BacklogTask, sortOrder: Int) {
+    init(from task: BacklogTask, sortOrder: Double) {
         self.taskId = task.id
         self.title = task.title
         self.durationMinutes = task.durationMinutes
